🐛fix: 컴포넌트 작성
