+++ src/components/writeRetro/task/ReviseModal.tsx
@@ -2,6 +2,7 @@ import { ChangeEvent, FC, useState } from 'react';
 import { Button } from '@chakra-ui/react';
 import { sectionData } from '@/api/@types/Section';
 import { SectionServices } from '@/api/services/Section';
+import { useCustomToast } from '@/hooks/useCustomToast';
 import * as S from '@/styles/writeRetroStyles/Layout.style';
 
 interface Props {
@@ -10,7 +11,9 @@ interface Props {
 
 const ReviseModal: FC<Props> = ({ section }) => {
   // Input 높이 자동 조절
-  const [value, setValue] = useState('');
+  const [value, setValue] = useState<string>('');
+  const toast = useCustomToast();
+
   const handleChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     setValue(e.target.value);
     e.target.style.height = 'auto';
@@ -19,10 +22,9 @@ const ReviseModal: FC<Props> = ({ section }) => {
 
   const ChangeContent = async () => {
     try {
-      const data = await SectionServices.patch({ sectionId: section.sectionId, sectionContent: value });
-      console.log(data);
+      await SectionServices.patch({ sectionId: section.sectionId, sectionContent: value });
     } catch (e) {
-      console.error(e);
+      toast.error(e);
     }
   };
   return (
